Moved SharedNumberSequence to the deprecated package
