+++ api-report/sequence.api.md
@@ -439,38 +439,6 @@ export class SequenceMaintenanceEvent extends SequenceEvent<MergeTreeMaintenance
     readonly opArgs: IMergeTreeDeltaOpArgs | undefined;
 }
 
-// @public @deprecated
-export class SharedNumberSequence extends SharedSequence<number> {
-    // @deprecated
-    constructor(document: IFluidDataStoreRuntime, id: string, attributes: IChannelAttributes);
-    // @deprecated
-    static create(runtime: IFluidDataStoreRuntime, id?: string): SharedNumberSequence;
-    // @deprecated
-    static getFactory(): SharedNumberSequenceFactory;
-    // @deprecated (undocumented)
-    getRange(start: number, end?: number): number[];
-    // (undocumented)
-    id: string;
-}
-
-// @public @deprecated (undocumented)
-export class SharedNumberSequenceFactory implements IChannelFactory {
-    // @deprecated (undocumented)
-    static readonly Attributes: IChannelAttributes;
-    // @deprecated (undocumented)
-    get attributes(): IChannelAttributes;
-    // @deprecated (undocumented)
-    create(document: IFluidDataStoreRuntime, id: string): ISharedObject;
-    // @deprecated (undocumented)
-    load(runtime: IFluidDataStoreRuntime, id: string, services: IChannelServices, attributes: IChannelAttributes): Promise<ISharedObject>;
-    // @deprecated (undocumented)
-    static segmentFromSpec(segSpec: IJSONSegment): SubSequence<number>;
-    // @deprecated (undocumented)
-    static Type: string;
-    // @deprecated (undocumented)
-    get type(): string;
-}
-
 // @public (undocumented)
 export abstract class SharedSegmentSequence<T extends ISegment> extends SharedObject<ISharedSegmentSequenceEvents> implements ISharedIntervalCollection<SequenceInterval> {
     constructor(dataStoreRuntime: IFluidDataStoreRuntime, id: string, attributes: IChannelAttributes, segmentFromSpec: (spec: IJSONSegment) => ISegment);
