+++ examples/data-objects/table-document/src/document.ts
@@ -11,11 +11,11 @@ import { ISequencedDocumentMessage } from "@fluidframework/protocol-definitions"
 import {
     positionToRowCol,
     rowColToPosition,
-    SharedNumberSequence,
     SparseMatrix,
     IntervalType,
     SequenceDeltaEvent,
 } from "@fluidframework/sequence";
+import { SharedNumberSequence } from "@fluid-experimental/deprecated";
 import { CellRange } from "./cellrange";
 import { TableDocumentType } from "./componentTypes";
 import { ConfigKey } from "./configKey";
