+++ experimental/dds/deprecated/src/test/sharedObjectSequence.spec.ts
@@ -10,7 +10,8 @@ import {
     MockFluidDataStoreRuntime,
     MockStorage,
 } from "@fluidframework/test-runtime-utils";
-import { SharedObjectSequence, SharedObjectSequenceFactory } from "../sharedObjectSequence";
+import { SharedObjectSequence } from "../sharedObjectSequence";
+import { SharedObjectSequenceFactory } from "../sequenceFactory";
 
 function createConnectedSequence(id: string, runtimeFactory: MockContainerRuntimeFactory) {
     const dataStoreRuntime = new MockFluidDataStoreRuntime();
