+++ packages/dds/sequence/src/index.ts
@@ -26,5 +26,4 @@ export * from "./sequence";
 export * from "./sequenceFactory";
 export * from "./sequenceDeltaEvent";
 export * from "./sharedSequence";
-export * from "./sharedNumberSequence";
 export * from "./sparsematrix";
