+++ packages/dds/sequence/src/test/types/validateSequencePrevious.ts
@@ -738,50 +738,26 @@ use_old_ClassDeclaration_SequenceMaintenanceEvent(
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_SharedNumberSequence": {"forwardCompat": false}
+* "RemovedClassDeclaration_SharedNumberSequence": {"forwardCompat": false}
 */
-declare function get_old_ClassDeclaration_SharedNumberSequence():
-    TypeOnly<old.SharedNumberSequence>;
-declare function use_current_ClassDeclaration_SharedNumberSequence(
-    use: TypeOnly<current.SharedNumberSequence>);
-use_current_ClassDeclaration_SharedNumberSequence(
-    get_old_ClassDeclaration_SharedNumberSequence());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_SharedNumberSequence": {"backCompat": false}
+* "RemovedClassDeclaration_SharedNumberSequence": {"backCompat": false}
 */
-declare function get_current_ClassDeclaration_SharedNumberSequence():
-    TypeOnly<current.SharedNumberSequence>;
-declare function use_old_ClassDeclaration_SharedNumberSequence(
-    use: TypeOnly<old.SharedNumberSequence>);
-use_old_ClassDeclaration_SharedNumberSequence(
-    get_current_ClassDeclaration_SharedNumberSequence());
 
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_SharedNumberSequenceFactory": {"forwardCompat": false}
+* "RemovedClassDeclaration_SharedNumberSequenceFactory": {"forwardCompat": false}
 */
-declare function get_old_ClassDeclaration_SharedNumberSequenceFactory():
-    TypeOnly<old.SharedNumberSequenceFactory>;
-declare function use_current_ClassDeclaration_SharedNumberSequenceFactory(
-    use: TypeOnly<current.SharedNumberSequenceFactory>);
-use_current_ClassDeclaration_SharedNumberSequenceFactory(
-    get_old_ClassDeclaration_SharedNumberSequenceFactory());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_SharedNumberSequenceFactory": {"backCompat": false}
+* "RemovedClassDeclaration_SharedNumberSequenceFactory": {"backCompat": false}
 */
-declare function get_current_ClassDeclaration_SharedNumberSequenceFactory():
-    TypeOnly<current.SharedNumberSequenceFactory>;
-declare function use_old_ClassDeclaration_SharedNumberSequenceFactory(
-    use: TypeOnly<old.SharedNumberSequenceFactory>);
-use_old_ClassDeclaration_SharedNumberSequenceFactory(
-    get_current_ClassDeclaration_SharedNumberSequenceFactory());
 
 /*
 * Validate forward compat by using old type in place of current type
